chapter_7/recipe_6/requirements.txt
azure-functions==1.7.2
requests==2.26.0
azure-keyvault-secrets==4.3.0
